+++ CHANGELOG.md
@@ -52,6 +52,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `LBFGS` as available optimizer option. LBFGS is a Quasi-Newton 2nd order optimizer that was designed for full-batch update scheme and converges very fast on certain convex problems. While generally it requires full-batch update and ideally noiseless gradients, it's included for completeness and certain special cases.
 - Add `.json` as supported params file type in hope of better support for LLM-related interfaces
+- Add `slice_thickness` as an AD-optimizable parameter and rename most `z_distance` into `slice_thickness` for clarity.
+- Add `plot_slice_thickness` for visualization
+### Changed
+- Refine `near_field_evolution` (Fresnel propagator) for performance and simplify the function signature.
+- Refactor `recon_step` and `recon_loop` a bit by moving `iter_times`, `loss_iters`, and `dz_iters` all inside `recon_step` since they'll be stored in `model` for every iteration
+- Simplify the function signature of `save_results`, `make_save_dict`, and `plot_summary` by moving the iter-result-lists inside `model`
 
 ## [v0.1.0-beta2.10] - 2024-11-06
 ### Added